rgwlib: rework linkage
