+++ src/test/librgw_file_nfsns.cc
@@ -21,6 +21,7 @@
 #include "include/rados/librgw.h"
 #include "include/rados/rgw_file.h"
 #include "rgw/rgw_file.h"
+#include "rgw_lib.h"
 #include "rgw/rgw_lib_frontend.h" // direct requests
 
 #include "gtest/gtest.h"
@@ -31,6 +32,7 @@
 #define dout_subsys ceph_subsys_rgw
 
 using namespace std;
+using namespace rgw; // g_rgwlib
 
 namespace {
 
@@ -259,9 +261,9 @@ TEST(LibRGW, SETUP_HIER1)
 		    << std::endl;
 	}
 	RGWPutObjRequest req(cct,
-			     rgwlib.get_store()->get_user(fs_private->get_user()->user_id),
+			     g_rgwlib->get_store()->get_user(fs_private->get_user()->user_id),
 			     bucket_name, obj_name, bl);
-	int rc = rgwlib.get_fe()->execute_req(&req);
+	int rc = g_rgwlib->get_fe()->execute_req(&req);
 	int rc2 = req.get_ret();
 	ASSERT_EQ(rc, 0);
 	ASSERT_EQ(rc2, 0);
